test: ensures that return 422 when request is invalid to CreateProjectControllerIT
