+++ infrastructure/src/test/java/com/nucleo42/infrastruture/controller/FindAllProjectControllerTest.java
@@ -1,7 +1,6 @@
 package com.nucleo42.infrastruture.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.nucleo42.entity.Project;
 import com.nucleo42.infrastruture.dto.CreateProjectRequestDTO;
 import com.nucleo42.infrastruture.entity.ProjectEntity;
 import com.nucleo42.infrastruture.repository.ProjectRepository;
@@ -13,14 +12,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.MediaType;
-import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
 
 import java.util.List;
-import java.util.UUID;
 
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
@@ -69,25 +66,6 @@ class FindAllProjectControllerTest {
                 )
         );
         projectEntityRepository.saveAll(projects);
-    }
-
-    @Test
-    @DisplayName("Should create a project and return status 201 on success")
-    void test01() throws Exception {
-        mockMvc.perform(post("/project")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(objectMapper.writeValueAsString(createProjectRequestDTO)))
-                .andExpect(status().isCreated());
-    }
-
-    @Test
-    @DisplayName("Should return 404 when request is invalid")
-    void test02() throws Exception {
-        createProjectRequestDTO = new CreateProjectRequestDTO("", "", 0, "");
-        mockMvc.perform(post("/project")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(objectMapper.writeValueAsString(createProjectRequestDTO)))
-                .andExpect(status().isBadRequest());
     }
 
     @Test
